test: keep reconstruction proof within service boundaries
